Fix TestRegistry and add tests

+++ Functions/TestRegistry.ps1
@@ -24,17 +24,31 @@ function New-TestRegistry
     }
 
     $DriveName = "TestRegistry"
-
     #setup the test drive
     if ( -not (& $SafeCommands['Test-Path'] "${DriveName}:\") )
     {
-        $null = & $SafeCommands['New-PSDrive'] -Name $DriveName -PSProvider Registry -Root $directory -Scope Global -Description "Pester test drive"
-    }
+        try {
+            $null = & $SafeCommands['New-PSDrive'] -Name $DriveName -PSProvider Registry -Root $directory -Scope Global -Description "Pester test registry" -ErrorAction Stop
+        }
+        catch {
+            if ($_.FullyQualifiedErrorId -like 'DriveAlreadyExists*') {
+                # it can happen that Test-Path reports false even though the drive
+                # exists. I don't know why but I see it in "Context Teardown fails"
+                # it would be possible to use Get-PsDrive directly for the test but it
+                # is about 10ms slower and we do it in every Describe and It so it would
+                # quickly add up
 
-    #publish the global TestRegistry variable used in few places within the module
-    if (-not (& $SafeCommands['Test-Path'] "Variable:Global:$DriveName"))
-    {
-        & $SafeCommands['New-Variable'] -Name $DriveName -Scope Global -Value $directory
+                # so if that happens just ignore the error, the goal of this function is to
+                # create the testdrive and the testdrive already exists, so all is good.
+            }
+            else
+            {
+                Write-Error $_ -ErrorAction 'Stop'
+            }
+        }
+    }
+    else {
+        write-host "registry drive exists"
     }
 
     if ( $PassThru )
@@ -43,6 +57,9 @@ function New-TestRegistry
     }
 }
 
+function Get-TestRegistryPath () {
+    "Microsoft.PowerShell.Core\Registry::" + (& $SafeCommands['Get-PSDrive'] -Name TestRegistry).Root
+}
 
 function Clear-TestRegistry
 {
@@ -51,24 +68,22 @@ function Clear-TestRegistry
         $Exclude
     )
 
-    $Path = (& $SafeCommands['Get-PSDrive'] -Name TestRegistry).Root
+    $path = Get-TestRegistryPath
 
-    if (& $SafeCommands['Test-Path'] -Path $Path )
+    if ($null -ne $path -and (& $SafeCommands['Test-Path'] -Path $Path))
     {
         #Get-ChildItem -Exclude did not seem to work with full paths
         & $SafeCommands['Get-ChildItem'] -Recurse -Path $Path |
-            & $SafeCommands['Sort-Object'] -Descending  -Property "FullName" |
-            & $SafeCommands['Where-Object'] { $Exclude -NotContains $_.FullName } |
+            & $SafeCommands['Sort-Object'] -Descending  -Property 'PSPath' |
+            & $SafeCommands['Where-Object'] { $Exclude -NotContains $_.PSPath } |
             & $SafeCommands['Remove-Item'] -Force -Recurse
     }
 }
 
 function Get-TestRegistryChildItem {
-    $Path = (& $SafeCommands['Get-PSDrive'] -Name TestRegistry).Root
-    if (& $SafeCommands['Test-Path'] -Path $Path )
-    {
-        & $SafeCommands['Get-ChildItem'] -Recurse -Path $Path
-    }
+    $path = Get-TestRegistryPath
+
+    & $SafeCommands['Get-ChildItem'] -Recurse -Path $path
 }
 
 function New-RandomTempRegistry
@@ -84,10 +99,16 @@ function New-RandomTempRegistry
 
 function Remove-TestRegistry
 {
-
     $DriveName = "TestRegistry"
     $Drive = & $SafeCommands['Get-PSDrive'] -Name $DriveName -ErrorAction $script:IgnoreErrorPreference
-    $Path = ($Drive).Root
+    if ($null -eq $Drive) {
+        # the drive does not exist, someone must have removed it instead of us,
+        # most likely a test that tests pester itself, so we just hope that the
+        # one who removed this removed also the contents of it correctly
+        return
+    }
+
+    $path = Get-TestRegistryPath
 
     if ($pwd -like "$DriveName*" )
     {
@@ -101,62 +122,13 @@ function Remove-TestRegistry
         $Drive | & $SafeCommands['Remove-PSDrive'] -Force #This should fail explicitly as it impacts future pester runs
     }
 
-    if (& $SafeCommands['Test-Path'] -Path $Path)
+    if (& $SafeCommands['Test-Path'] -Path $path)
     {
-        & $SafeCommands['Remove-Item'] -Path $Path -Force -Recurse
+        & $SafeCommands['Remove-Item'] -Path $path -Force -Recurse
     }
 
     if (& $SafeCommands['Get-Variable'] -Name $DriveName -Scope Global -ErrorAction $script:IgnoreErrorPreference)
     {
         & $SafeCommands['Remove-Variable'] -Scope Global -Name $DriveName -Force
-    }
-}
-
-function Initialize-TestRegistry
-{
-    param(
-        [Parameter(Mandatory = $true)]
-        [ValidateSet("Key", "Property")]
-        $ObjectType,
-
-        [Parameter(Mandatory = $true)]
-        [System.String]
-        $Path,
-
-        [System.String]
-        $PropertyName,
-
-        [ValidateSet("String", "ExpandString", "Binary", "DWord", "MultiString", "Qword", "Unknown")]
-        $PropertyType,
-
-        $Value,
-
-        [switch]
-        $PassThru
-    )
-
-    Assert-DescribeInProgress -CommandName Setup
-
-    $TestRegistryName = & $SafeCommands['Get-PSDrive'] TestRegistry |
-        & $SafeCommands['Select-Object'] -ExpandProperty Root
-
-    $Path = $Path.Replace(":", "")
-
-    switch ($ObjectType)
-    {
-        "Key"
-        {
-            $item = & $SafeCommands['New-Item'] -Path "${TestRegistryName}\${Path}" -Value $Value -Force
-        }
-        "Property"
-        {
-            $item = & $SafeCommands['New-ItemProperty'] -Name $PropertyName -Path "${TestRegistryName}\${Path}" -Value $Value -PropertyType $PropertyType -Force
-        }
-
-    }
-
-    if ($PassThru)
-    {
-        return $item
     }
 }
